#24 Introducing minimum fee calculation. Added documentation and tests.
