+++ apps/utils/lib/governance.ex
@@ -1,4 +1,7 @@
 defmodule Utils.Governance do
+  @moduledoc """
+  Contains all constants and helper functions, related to blockchain.
+  """
   alias AeternityNode.Model.{
     SpendTx,
     OracleRegisterTx,
@@ -23,25 +26,14 @@ defmodule Utils.Governance do
     ContractCreateTx
   }
 
-  @tx_base_gas 15_000
   @byte_gas 20
-  # 32000 as `GCREATE` i.e. an oracle-related state object costs per year as much as it costs to indefinitely create an account.
+  @tx_base_gas 15_000
+  # 32_000 as `GCREATE` i.e. an oracle-related state object costs per year as much as it costs to indefinitely create an account.
   @oracle_state_gas_per_year 32_000
   @expected_block_mine_rate_minutes 3
   @expected_blocks_in_a_year_floor 175_200 = div(60 * 24 * 365, @expected_block_mine_rate_minutes)
 
-  def tx_base_gas(%ContractCallTx{}), do: 30 * @tx_base_gas
-  def tx_base_gas(%ContractCreateTx{}), do: 5 * @tx_base_gas
   def tx_base_gas(%SpendTx{}), do: @tx_base_gas
-  def tx_base_gas(%ChannelDepositTx{}), do: @tx_base_gas
-  def tx_base_gas(%ChannelCloseMutualTx{}), do: @tx_base_gas
-  def tx_base_gas(%ChannelCloseSoloTx{}), do: @tx_base_gas
-  def tx_base_gas(%ChannelCreateTx{}), do: @tx_base_gas
-  def tx_base_gas(%ChannelForceProgressTx{}), do: @tx_base_gas
-  def tx_base_gas(%ChannelSlashTx{}), do: @tx_base_gas
-  def tx_base_gas(%ChannelSettleTx{}), do: @tx_base_gas
-  def tx_base_gas(%ChannelSnapshotSoloTx{}), do: @tx_base_gas
-  def tx_base_gas(%ChannelWithdrawTx{}), do: @tx_base_gas
   def tx_base_gas(%NamePreclaimTx{}), do: @tx_base_gas
   def tx_base_gas(%NameClaimTx{}), do: @tx_base_gas
   def tx_base_gas(%NameTransferTx{}), do: @tx_base_gas
@@ -51,15 +43,33 @@ defmodule Utils.Governance do
   def tx_base_gas(%OracleQueryTx{}), do: @tx_base_gas
   def tx_base_gas(%OracleRespondTx{}), do: @tx_base_gas
   def tx_base_gas(%OracleExtendTx{}), do: @tx_base_gas
+  def tx_base_gas(%ContractCallTx{}), do: 30 * @tx_base_gas
+  def tx_base_gas(%ContractCreateTx{}), do: 5 * @tx_base_gas
+  def tx_base_gas(%ChannelDepositTx{}), do: @tx_base_gas
+  def tx_base_gas(%ChannelCloseMutualTx{}), do: @tx_base_gas
+  def tx_base_gas(%ChannelCloseSoloTx{}), do: @tx_base_gas
+  def tx_base_gas(%ChannelCreateTx{}), do: @tx_base_gas
+  def tx_base_gas(%ChannelForceProgressTx{}), do: @tx_base_gas
+  def tx_base_gas(%ChannelSlashTx{}), do: @tx_base_gas
+  def tx_base_gas(%ChannelSettleTx{}), do: @tx_base_gas
+  def tx_base_gas(%ChannelSnapshotSoloTx{}), do: @tx_base_gas
+  def tx_base_gas(%ChannelWithdrawTx{}), do: @tx_base_gas
+
+  @spec byte_gas() :: non_neg_integer()
   def byte_gas(), do: @byte_gas
 
-  def state_gas_per_block(%OracleRegisterTx{}),
-    do: {@oracle_state_gas_per_year, @expected_blocks_in_a_year_floor}
+  @spec state_gas_per_block(
+          OracleRegisterTx.t()
+          | OracleRespondTx.t()
+          | OracleQueryTx.t()
+          | OracleExtendTx.t()
+        ) :: {non_neg_integer(), non_neg_integer()}
+  def state_gas_per_block(%struct{})
+      when struct in [OracleRegisterTx, OracleRespondTx, OracleQueryTx, OracleExtendTx] do
+    {@oracle_state_gas_per_year, @expected_blocks_in_a_year_floor}
+  end
 
-  def state_gas_per_block(%OracleQueryTx{}), do: state_gas_per_block(%OracleRegisterTx{})
-  def state_gas_per_block(%OracleRespondTx{}), do: state_gas_per_block(%OracleRegisterTx{})
-  def state_gas_per_block(%OracleExtendTx{}), do: state_gas_per_block(%OracleRegisterTx{})
-
+  @spec state_gas(tuple(), non_neg_integer()) :: non_neg_integer()
   def state_gas({part, whole}, n_key_blocks)
       when is_integer(whole) and whole > 0 and is_integer(part) and part >= 0 and
              is_integer(n_key_blocks) and n_key_blocks >= 0 do
@@ -67,7 +77,7 @@ defmodule Utils.Governance do
     div(tmp + (whole - 1), whole)
   end
 
-  @spec min_gas_price(non_neg_integer(), String.t()) :: 1 | 1_000_000
+  @spec min_gas_price(non_neg_integer(), String.t()) :: non_neg_integer()
   def min_gas_price(height, network_id)
       when is_integer(height) and height >= 0 and is_binary(network_id) do
     case protocol_effective_at_height(height, network_id) do
@@ -77,10 +87,7 @@ defmodule Utils.Governance do
   end
 
   defp protocol_effective_at_height(height, "ae_mainnet") when height < 47_800, do: 1
-
   defp protocol_effective_at_height(height, "ae_mainnet") when height >= 47_800, do: 2
-
   defp protocol_effective_at_height(height, "ae_uat") when height < 40_900, do: 1
-
   defp protocol_effective_at_height(height, "ae_uat") when height >= 40_900, do: 2
 end
